pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "polintel-feeds"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.2",
    "playwright>=1.44",
    "anthropic>=0.28",
    "feedgen>=1.0",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
polintel = "src.cli:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
